refactor : 유저 응답 코드 폴더 위치 변경
